UI to query completed esign packages for a 24 hour period
Complete site to query completed packages. It allows the user to select
a start date and time. Validation exists to prevent users from inputting
incorrect values. A query will be made to an azure function to retrieve
the completed packages. Upon response a list of packages and their
documents will be displayed, along with a total document count.